feat: create payslip

+++ src/main/java/com/example/payslip/data/entity/EmployeeEntity.java
@@ -6,6 +6,7 @@ import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.UUID;
 
@@ -19,6 +20,6 @@ public class EmployeeEntity {
     private UUID id;
     private String username;
     private String password;
-    private BigInteger salary;
+    private BigDecimal salary;
 
 }
