+++ src/main/resources/db/migration/V1__create-employees-table.sql
@@ -6,7 +6,7 @@ CREATE TABLE employees(
     id  UUID NOT NULL DEFAULT gen_random_uuid() PRIMARY KEY,
     username VARCHAR(100) NOT NULL,
     password TEXT NOT NULL,
-    salary BIGINT NOT NULL,
+    salary DECIMAL NOT NULL,
     CONSTRAINT salary_range CHECK (salary > 0)
 );
